Fix field_reader to handle boolean defaults
We only want to set a default on a field if there is no value currently
set in Contentful. In its previous implementation we wouldn't be able to
set a default value of "true" for a boolean field because a value of
"false" would trigger the default. This changes `field_reader` to check
to see if the value is explicitly `nil` before falling back to the
default.